requirements.txt
bs4==0.0.1
fake-useragent==1.1.0
finance-datareader==0.9.50
gensim==4.2.0
graphviz==0.20.1
joblib==1.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
konlpy==0.6.0
kss==3.6.4
lightgbm==3.3.3
lxml==4.8.0
matplotlib==3.2.2
matplotlib-inline==0.1.3
mplfinance==0.12.9b0
mysqlclient==2.1.1
nltk==3.7
numpy==1.23.4
pandas==1.4.1
pandas-datareader==0.10.0
prophet==1.1.1
pydot==1.4.2
pykrx==1.0.33
PyMySQL==1.0.2
pystan==2.19.1.1
python-mecab-kor==1.2.3
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
Scrapy==2.6.2
scrapy-fake-useragent==1.4.4
seaborn==0.11.2
selenium==4.4.3
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.2
tensorflow-addons==0.18.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.27.0
tokenizers==0.13.2
tqdm==4.64.0
transformers==4.24.0
wordcloud==1.8.2.2
xgboost==1.6.2
yfinance==0.1.70
